update: move the delete function from postCard component to post hooks functions to make the code in modular structure
